+++ react/components/ListDetails/Footer.tsx
@@ -3,9 +3,8 @@ import React, { Component, ReactNode } from 'react'
 import { map, path, head } from 'ramda'
 import { compose } from 'react-apollo'
 import { isMobile } from 'react-device-detect'
-import { FormattedMessage, InjectedIntlProps, injectIntl } from 'react-intl'
+import { FormattedMessage, InjectedIntlProps, injectIntl, defineMessages } from 'react-intl'
 import { withToast } from 'vtex.styleguide'
-import messages from '../../constants/messages'
 
 import BuyButton from 'vtex.store-components/BuyButton'
 import ProductPrice from 'vtex.store-components/ProductPrice'
@@ -19,6 +18,21 @@ interface FooterProps extends InjectedIntlProps {
 interface FooterState {
   isLoading?: boolean
 }
+
+const messages = defineMessages({
+  selectedItemsQuantity: {
+    defaultMessage: '',
+    id: 'wishlist-quantity-selected-items',
+  },
+  total: {
+    defaultMessage: '',
+    id: 'wishlist-total',
+  },
+  buyItems: {
+    defaultMessage: '',
+    id: 'wishlist-buy-items',
+  },
+})
 
 class Footer extends Component<FooterProps, FooterState> {
   public state: FooterState = {
