--- react/constants/messages.ts
@@ -1,130 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-const messages = defineMessages({
-  seeLists: {
-    defaultMessage: '',
-    id: 'wishlist-see-lists',
-  },
-  productAddedToList: {
-    defaultMessage: '',
-    id: 'wishlist-product-added-to-list',
-  },
-  addProductFail: {
-    defaultMessage: '',
-    id: 'wishlist-add-product-fail',
-  },
-  listNameDefault: {
-    defaultMessage: '',
-    id: 'wishlist-default-list-name',
-  },
-  cancelDialog: {
-    defaultMessage: '',
-    id: 'wishlist-dialog-cancel',
-  },
-  confirmDialog: {
-    defaultMessage: '',
-    id: 'wishlist-dialog-confirm',
-  },
-  optionConfiguration: {
-    defaultMessage: '',
-    id: 'wishlist-option-configuration',
-  },
-  optionDelete: {
-    defaultMessage: '',
-    id: 'wishlist-option-delete',
-  },
-  messageDeleteConfirmation: {
-    defaultMessage: '',
-    id: 'wishlist-delete-confirmation-message',
-  },
-  addToList: {
-    defaultMessage: '',
-    id: 'wishlist-add-to-list',
-  },
-  apply: {
-    defaultMessage: '',
-    id: 'wishlist-apply',
-  },
-  new: {
-    defaultMessage: '',
-    id: 'wishlist-new',
-  },
-  addButton: {
-    defaultMessage: '',
-    id: 'wishlist-add-button',
-  },
-  listNamePlaceholder: {
-    defaultMessage: '',
-    id: 'wishlist-list-name-placeholder',
-  },
-  listNameLabel: {
-    defaultMessage: '',
-    id: 'wishlist-list-name-label',
-  },
-  isPublic: {
-    defaultMessage: '',
-    id: 'wishlist-is-public',
-  },
-  isPublicHint: {
-    defaultMessage: '',
-    id: 'wishlist-is-public-hint',
-  },
-  save: {
-    defaultMessage: '',
-    id: 'wishlist-save',
-  },
-  listUpdate: {
-    defaultMessage: '',
-    id: 'wishlist-list-updated',
-  },
-  copyTo: {
-    defaultMessage: '',
-    id: 'wishlist-copy-to',
-  },
-  see: {
-    defaultMessage: '',
-    id: 'wishlist-see',
-  },
-  copied: {
-    defaultMessage: '',
-    id: 'wishlist-copied',
-  },
-  myLists: {
-    defaultMessage: '',
-    id: 'wishlist-my-lists',
-  },
-  noListCreated: {
-    defaultMessage: '',
-    id: 'wishlist-no-list-created',
-  },
-  productsQuantity: {
-    defaultMessage: '',
-    id: 'wishlist-quantity-products',
-  },
-  itemsQuantity: {
-    defaultMessage: '',
-    id: 'wishlist-quantity-of-items',
-  },
-  emptyList: {
-    defaultMessage: '',
-    id: 'wishlist-list-empty',
-  },
-  addItems: {
-    defaultMessage: '',
-    id: 'wishlist-add-items',
-  },
-  selectedItemsQuantity: {
-    defaultMessage: '',
-    id: 'wishlist-quantity-selected-items',
-  },
-  total: {
-    defaultMessage: '',
-    id: 'wishlist-total',
-  },
-  buyItems: {
-    defaultMessage: '',
-    id: 'wishlist-buy-items',
-  },
-})
-
-export default messages
